--- src/main/kotlin/net/battaglini/fantaf1appbackend/controller/AdminOperationsController.kt
@@ -1,113 +1,0 @@
-package net.battaglini.fantaf1appbackend.controller
-
-import net.battaglini.fantaf1appbackend.exception.InvalidRequestException
-import net.battaglini.fantaf1appbackend.model.request.CalculateTeamsResultsRequest
-import net.battaglini.fantaf1appbackend.model.request.GenerateRaceRecapRequest
-import net.battaglini.fantaf1appbackend.model.request.UpdateDriversPricesRequest
-import net.battaglini.fantaf1appbackend.model.request.UpdateDriversSummariesRequest
-import net.battaglini.fantaf1appbackend.model.response.CalculateTeamsResultsResponse
-import net.battaglini.fantaf1appbackend.model.response.DriverPriceUpdateResponse
-import net.battaglini.fantaf1appbackend.model.response.DriverSummariesResponse
-import net.battaglini.fantaf1appbackend.model.response.GenerateRaceRecapResponse
-import net.battaglini.fantaf1appbackend.service.DriverPricingService
-import net.battaglini.fantaf1appbackend.service.DriverService
-import net.battaglini.fantaf1appbackend.service.RaceWeekendService
-import net.battaglini.fantaf1appbackend.service.TeamResultsService
-import org.slf4j.LoggerFactory
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping(path = ["/admin"])
-class AdminOperationsController(
-    private val driverService: DriverService,
-    private val driverPricingService: DriverPricingService,
-    private val raceWeekendService: RaceWeekendService,
-    private val teamResultsService: TeamResultsService
-) {
-    @PostMapping("/drivers/seed")
-    suspend fun seedDrivers() {
-        try {
-            driverService.seedDrivers()
-        } catch (e: Exception) {
-            throw RuntimeException(e.message)
-        }
-    }
-
-    @PostMapping("/race-weekends/seed")
-    suspend fun seedRaceWeekends() {
-        try {
-            raceWeekendService.seedRaceWeekends()
-        } catch (e: Exception) {
-            throw RuntimeException(e.message)
-        }
-    }
-
-    @PostMapping("/drivers/summaries")
-    suspend fun updateDriversSummaries(@RequestBody body: UpdateDriversSummariesRequest): DriverSummariesResponse {
-        val summaries = driverService.updateDriverSummaries(body.acronyms)
-        return DriverSummariesResponse(summaries)
-    }
-
-    @PostMapping("/drivers/prices")
-    suspend fun updateDriversPrices(@RequestBody request: UpdateDriversPricesRequest): DriverPriceUpdateResponse {
-        try {
-            val updates = driverPricingService.calculateAndUpdatePrices(
-                acronyms = request.acronyms,
-                updateAll = request.updateAllDrivers
-            )
-            return DriverPriceUpdateResponse(updates)
-        } catch (e: Exception) {
-            LOGGER.error("Failed to manually update driver prices", e)
-            throw RuntimeException(e.message)
-        }
-    }
-
-    @PostMapping("/teams/results")
-    suspend fun calculateTeamsResults(@RequestBody request: CalculateTeamsResultsRequest): CalculateTeamsResultsResponse {
-        try {
-            val raceWeekend = raceWeekendService.getRaceWeekend(request.raceId)
-                ?: throw InvalidRequestException("RaceWeekend with raceId=${request.raceId} not found")
-            val raceWeekendResults = raceWeekendService.getRaceWeekendResults(request.raceId)
-                ?: throw InvalidRequestException("No results found for raceId=${request.raceId}")
-
-            val teamsResults = teamResultsService.calculateAndSaveLineupsResults(raceWeekendResults)
-
-            return CalculateTeamsResultsResponse(
-                raceId = raceWeekend.raceId,
-                raceName = raceWeekend.raceName,
-                scores = teamsResults.map { CalculateTeamsResultsResponse.Companion.TeamScore(it.key, it.value) }
-            )
-        } catch (e: Exception) {
-            LOGGER.error("Failed to calculate teams results for raceId={}", request.raceId, e)
-            throw RuntimeException(e.message)
-        }
-    }
-
-    @PostMapping("/race-weekends/recap")
-    suspend fun generateRaceRecaps(@RequestBody request: GenerateRaceRecapRequest): GenerateRaceRecapResponse {
-        try {
-        val recaps = raceWeekendService.generateRaceRecap(request.raceIds)
-            val recapEntries = recaps.map {
-                GenerateRaceRecapResponse.RecapEntry(
-                    raceId = it.raceId,
-                    raceName = it.raceName,
-                    recapParagraphs = it.recapParagraphs
-                )
-            }
-            return GenerateRaceRecapResponse(
-                recapIds = recaps.map { it.raceId },
-                recaps = recapEntries
-            )
-        } catch (e: Exception) {
-            LOGGER.error("Failed to generate race recaps", e)
-            throw RuntimeException(e.message)
-        }
-    }
-
-    companion object {
-        private val LOGGER = LoggerFactory.getLogger(AdminOperationsController::class.java)
-    }
-}
